fix BinaryOp sliced input
